refactor: used DocumentFragment
